fix(AntColony): Improve how we found solutions

+++ src/main/java/com/deliverif/app/algorithm/astar/GetLengthScorer.java
@@ -4,7 +4,7 @@ import com.deliverif.app.model.Intersection;
 
 public class GetLengthScorer implements Scorer<Intersection> {
     @Override
-    public double computeCost(Intersection from, Intersection to) {
+    public float computeCost(Intersection from, Intersection to) {
         return from.getReachableIntersections().get(to).getLength();
     }
 }
